test: fix tests on freebsd

+++ nofile_test_posix.go
@@ -7,5 +7,6 @@ import "syscall"
 func curFileLimit() uint64 {
 	var n syscall.Rlimit
 	syscall.Getrlimit(syscall.RLIMIT_NOFILE, &n)
-	return n.Cur
+	// cast because some platforms use int64 (e.g., freebsd)
+	return uint64(n.Cur)
 }
